Refactor reader & writer to separate files
